Group primitive classes to get better overview
