+++ src/som/interpreter/nodes/MessageSendNode.java
@@ -30,21 +30,6 @@ import som.interpreter.nodes.specialized.OrMessageNodeGen.OrBoolMessageNodeGen;
 import som.interpreter.nodes.specialized.whileloops.WhileWithDynamicBlocksNode;
 import som.interpreter.nodes.specialized.whileloops.WhileWithStaticBlocksNode.WhileFalseStaticBlocksNode;
 import som.interpreter.nodes.specialized.whileloops.WhileWithStaticBlocksNode.WhileTrueStaticBlocksNode;
-import som.primitives.BlockPrimsFactory.ValueNonePrimFactory;
-import som.primitives.BlockPrimsFactory.ValueOnePrimFactory;
-import som.primitives.EqualsEqualsPrimFactory;
-import som.primitives.EqualsPrimFactory;
-import som.primitives.IntegerPrimsFactory.AbsPrimFactory;
-import som.primitives.IntegerPrimsFactory.LeftShiftPrimFactory;
-import som.primitives.IntegerPrimsFactory.ToPrimFactory;
-import som.primitives.IntegerPrimsFactory.UnsignedRightShiftPrimFactory;
-import som.primitives.LengthPrimFactory;
-import som.primitives.MethodPrimsFactory.InvokeOnPrimFactory;
-import som.primitives.ObjectPrimsFactory.InstVarAtPrimFactory;
-import som.primitives.ObjectPrimsFactory.InstVarAtPutPrimFactory;
-import som.primitives.ObjectPrimsFactory.IsNilNodeGen;
-import som.primitives.ObjectPrimsFactory.NotNilNodeGen;
-import som.primitives.UnequalsPrimFactory;
 import som.primitives.arithmetic.AdditionPrimFactory;
 import som.primitives.arithmetic.BitXorPrimFactory;
 import som.primitives.arithmetic.DividePrimFactory;
@@ -64,6 +49,21 @@ import som.primitives.arrays.DoPrimFactory;
 import som.primitives.arrays.NewPrimFactory;
 import som.primitives.arrays.PutAllNodeFactory;
 import som.primitives.arrays.ToArgumentsArrayNodeGen;
+import som.primitives.basics.BlockPrimsFactory.ValueNonePrimFactory;
+import som.primitives.basics.BlockPrimsFactory.ValueOnePrimFactory;
+import som.primitives.basics.EqualsEqualsPrimFactory;
+import som.primitives.basics.EqualsPrimFactory;
+import som.primitives.basics.IntegerPrimsFactory.AbsPrimFactory;
+import som.primitives.basics.IntegerPrimsFactory.LeftShiftPrimFactory;
+import som.primitives.basics.IntegerPrimsFactory.ToPrimFactory;
+import som.primitives.basics.IntegerPrimsFactory.UnsignedRightShiftPrimFactory;
+import som.primitives.basics.LengthPrimFactory;
+import som.primitives.basics.UnequalsPrimFactory;
+import som.primitives.reflection.MethodPrimsFactory.InvokeOnPrimFactory;
+import som.primitives.reflection.ObjectPrimsFactory.InstVarAtPrimFactory;
+import som.primitives.reflection.ObjectPrimsFactory.InstVarAtPutPrimFactory;
+import som.primitives.reflection.ObjectPrimsFactory.IsNilNodeGen;
+import som.primitives.reflection.ObjectPrimsFactory.NotNilNodeGen;
 import som.vm.NotYetImplementedException;
 import som.vm.Universe;
 import som.vmobjects.SArray;
