+++ src/som/primitives/Primitives.java
@@ -33,9 +33,10 @@ import som.interpreter.Primitive;
 import som.interpreter.SomLanguage;
 import som.interpreter.nodes.ArgumentReadNode.LocalArgumentReadNode;
 import som.interpreter.nodes.ExpressionNode;
-import som.primitives.MethodPrimsFactory.InvokeOnPrimFactory;
 import som.primitives.arrays.PutAllNodeFactory;
 import som.primitives.arrays.ToArgumentsArrayNodeGen;
+import som.primitives.basics.LengthPrimFactory;
+import som.primitives.reflection.MethodPrimsFactory.InvokeOnPrimFactory;
 import som.vm.Universe;
 import som.vmobjects.SClass;
 import som.vmobjects.SInvokable;
