+++ src/som/primitives/basics/DoublePrims.java
@@ -1,10 +1,10 @@
-package som.primitives;
+package som.primitives.basics;
 
 import com.oracle.truffle.api.dsl.GenerateNodeFactory;
 import com.oracle.truffle.api.dsl.Specialization;
 
 import som.interpreter.nodes.nary.UnaryExpressionNode;
-import som.primitives.SystemPrims.UnarySystemNode;
+import som.primitives.basics.SystemPrims.UnarySystemNode;
 import som.vm.Universe;
 import som.vmobjects.SClass;
