+++ src/som/primitives/basics/UnequalsPrim.java
@@ -1,4 +1,4 @@
-package som.primitives;
+package som.primitives.basics;
 
 import java.math.BigInteger;
 
@@ -12,55 +12,55 @@ import som.vmobjects.SSymbol;
 
 
 @GenerateNodeFactory
-public abstract class EqualsPrim extends BinaryExpressionNode {
+public abstract class UnequalsPrim extends BinaryExpressionNode {
 
   private final SObject trueObject;
   private final SObject falseObject;
 
-  public EqualsPrim(final Universe universe) {
+  public UnequalsPrim(final Universe universe) {
     this.trueObject = universe.getTrueObject();
     this.falseObject = universe.getFalseObject();
   }
 
   @Specialization
   public final boolean doBoolean(final boolean left, final boolean right) {
-    return left == right;
+    return left != right;
   }
 
   @Specialization
   public final boolean doBoolean(final boolean left, final SObject right) {
-    return (left && right == trueObject) ||
-        (!left && right == falseObject);
+    return (left && right != trueObject) ||
+        (!left && right != falseObject);
   }
 
   @Specialization
   public final boolean doLong(final long left, final long right) {
-    return left == right;
+    return left != right;
   }
 
   @Specialization
   public final boolean doBigInteger(final BigInteger left, final BigInteger right) {
-    return left.compareTo(right) == 0;
+    return left.compareTo(right) != 0;
   }
 
   @Specialization
   public final boolean doString(final String receiver, final String argument) {
-    return receiver.equals(argument);
+    return !receiver.equals(argument);
   }
 
   @Specialization
   public final boolean doDouble(final double left, final double right) {
-    return left == right;
+    return left != right;
   }
 
   @Specialization
   public final boolean doSSymbol(final SSymbol left, final SSymbol right) {
-    return left == right;
+    return left != right;
   }
 
   @Specialization
   public final boolean doLong(final long left, final double right) {
-    return left == right;
+    return left != right;
   }
 
   @Specialization
@@ -79,47 +79,37 @@ public abstract class EqualsPrim extends BinaryExpressionNode {
   }
 
   @Specialization
-  public final boolean doString(final String receiver, final SSymbol argument) {
-    return receiver.equals(argument.getString());
-  }
-
-  @Specialization
-  public final boolean doSSymbol(final SSymbol receiver, final String argument) {
-    return receiver.getString().equals(argument);
-  }
-
-  @Specialization
   public final boolean doLong(final long left, final String right) {
-    return false;
+    return true;
   }
 
   @Specialization
   public final boolean doLong(final long left, final SObject right) {
-    return false;
+    return true;
   }
 
   @Specialization
   public final boolean doLong(final long left, final SSymbol right) {
-    return false;
+    return true;
   }
 
   @Specialization
   public final boolean doString(final String receiver, final long argument) {
-    return false;
+    return true;
   }
 
   @Specialization
   public final boolean doString(final String receiver, final SObject argument) {
-    return false;
+    return true;
   }
 
   @Specialization
   public final boolean doSSymbol(final SSymbol receiver, final long argument) {
-    return false;
+    return true;
   }
 
   @Specialization
   public final boolean doSSymbol(final SSymbol receiver, final SObject argument) {
-    return false;
+    return true;
   }
 }
