+++ src/som/primitives/reflection/GlobalPrim.java
@@ -1,4 +1,4 @@
-package som.primitives;
+package som.primitives.reflection;
 
 import com.oracle.truffle.api.dsl.Specialization;
 import com.oracle.truffle.api.frame.VirtualFrame;
@@ -6,8 +6,9 @@ import com.oracle.truffle.api.frame.VirtualFrame;
 import som.interpreter.nodes.ExpressionNode;
 import som.interpreter.nodes.GlobalNode;
 import som.interpreter.nodes.GlobalNode.UninitializedGlobalReadWithoutErrorNode;
+import som.primitives.basics.SystemPrims;
+import som.primitives.basics.SystemPrims.BinarySystemNode;
 import som.interpreter.nodes.SOMNode;
-import som.primitives.SystemPrims.BinarySystemNode;
 import som.vm.NotYetImplementedException;
 import som.vm.Universe;
 import som.vm.constants.Nil;
